+++ cypress/integration/performance.test.js
@@ -12,6 +12,7 @@ const { pages } = cypress
 const paths = [
   pages.pdp,
   pages.home,
+  pages.cmsHome,
   pages.collection,
   pages.collection_filtered,
 ]
@@ -80,6 +81,7 @@ describe('React rendering performance', () => {
 
     testPage(pages.pdp)
     testPage(pages.home)
+    testPage(pages.cmsHome)
     testPage(pages.search)
     testPage(pages.collection)
     testPage(pages.collection_filtered)