+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Alternative homepage rendered with CMS at `/cms` ([#104](https://github.com/vtex-sites/gatsby.store/pull/104))
+- Global Alert rendered using CMS ([#104](https://github.com/vtex-sites/gatsby.store/pull/104))
+- Integrates store with CMS ([#104](https://github.com/vtex-sites/gatsby.store/pull/104))
 - Applies new local tokens to `ProductGrid` ([#127](https://github.com/vtex-sites/gatsby.store/pull/127))
 - Applies new local tokens to `Tiles` ([#120](https://github.com/vtex-sites/gatsby.store/pull/120))
 - Applies new local tokens to `Accordion` ([#126](https://github.com/vtex-sites/gatsby.store/pull/126))